Move http status enum to infra layer
